firsts funcions checked

+++ ft_strlcat.c
@@ -6,11 +6,11 @@
 /*   By: bcanals- <bcanals-@student.42barcel>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/03 13:53:34 by bcanals-          #+#    #+#             */
-/*   Updated: 2024/06/27 02:19:24 by bizcru           ###   ########.fr       */
+/*   Updated: 2024/06/27 21:14:09 by bizcru           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-unsigned int	ft_strlcat(char *dest, const char *src, unsigned int size)
+long unsigned int	ft_strlcat(char *dest, const char *src, unsigned int size)
 {
 	unsigned int	i;
 	unsigned int	j;
